Adding Zokrates Verification Contracts
Again, using GitHub Co-Pilot to drive the initial structure of the contract and then filling the gaps and resolving the errors.

Verifier.sol was produced by Zokrates.

SolnSquareVerifier.sol will be the contract used by OpenSea as it has a reference to the ERC721Mintable contract and has access to it's functions.